+++ ConsoleApp1/Properties/AssemblyInfo.cs
