+++ TextAdventure.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26730.3
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sourcecode", "ConsoleApp1\Sourcecode.csproj", "{DC1AA5B2-6A27-44D7-85A7-3AB7D1D35A0C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DC1AA5B2-6A27-44D7-85A7-3AB7D1D35A0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DC1AA5B2-6A27-44D7-85A7-3AB7D1D35A0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DC1AA5B2-6A27-44D7-85A7-3AB7D1D35A0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DC1AA5B2-6A27-44D7-85A7-3AB7D1D35A0C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8A500C5A-4C5A-4981-BC80-2BACB4968996}
+	EndGlobalSection
+EndGlobal
